requirements.txt
transformers>=4.45
accelerate>=0.34
datasets>=3.0.0
evaluate>=0.4.1
numpy>=1.24
pandas>=2.0
tqdm>=4.66
pyyaml>=6.0
fire>=0.6.0
regex>=2024.5.15